Update to client v9200.2.0 (#139)
* Update to client v9200.1.0

* Update to v9200.2.0 to include float64 change

* Remove conversions from float32 to float64

* Fix accidentally removing config validators

+++ internal/acctest/resource/config/default_consent_service_resource_test.go
@@ -134,7 +134,7 @@ func testAccCheckExpectedConsentServiceAttributes(config consentServiceTestModel
 			return err
 		}
 		err = acctest.TestAttributesMatchInt(resourceType, nil, "search-size-limit",
-			config.search_size_limit, int64(*response.SearchSizeLimit))
+			config.search_size_limit, *response.SearchSizeLimit)
 		if err != nil {
 			return err
 		}
